+++ src/main/java/com/ic/invoicecapture/model/Debt.java
@@ -9,68 +9,6 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 public class Debt implements IModel, IRoutable {
-  public class Item implements IModel {
-
-    private String description;
-    private String name;
-    private Double price;
-    private Double quantity;
-    private Double vat;
-
-    public String getDescription() {
-      return description;
-    }
-
-    public String getName() {
-      return name;
-    }
-
-    public Double getPrice() {
-      return price;
-    }
-
-    public Double getQuantity() {
-      return quantity;
-    }
-
-    public Double getVat() {
-      return vat;
-    }
-
-    public void setDescription(String description) {
-      this.description = description;
-    }
-
-    public void setName(String name) {
-      this.name = name;
-    }
-
-    public void setPrice(Double price) {
-      this.price = price;
-    }
-
-    public void setQuantity(Double quantity) {
-      this.quantity = quantity;
-    }
-
-    public void setVat(Double vat) {
-      this.vat = vat;
-    }
-
-    @Override
-    public EnumMap<ItemField, Object> toEnumMap() {
-      EnumMap<ItemField, Object> map = new EnumMap<>(ItemField.class);
-
-      ModelUtils.tryAddObject(map, ItemField.NAME, getName());
-      ModelUtils.tryAddObject(map, ItemField.DESCRIPTION, getDescription());
-      ModelUtils.tryAddObject(map, ItemField.QUANTITY, getQuantity());
-      ModelUtils.tryAddObject(map, ItemField.VAT, getVat());
-      ModelUtils.tryAddObject(map, ItemField.PRICE, getPrice());
-
-      return map;
-    }
-  }
-
   private Map<String, String> attributes;
   private String currency;
   private String customerId;
@@ -110,11 +48,11 @@ public class Debt implements IModel, IRoutable {
   }
 
   public Date getDate() {
-    return date;
+    return new Date(date.getTime());
   }
 
   public Date getDueDate() {
-    return dueDate;
+    return new Date(dueDate.getTime());
   }
 
   public Double getGrossTotal() {
@@ -162,11 +100,11 @@ public class Debt implements IModel, IRoutable {
   }
 
   public void setDate(Date date) {
-    this.date = date;
+    this.date = new Date(date.getTime());
   }
 
   public void setDueDate(Date dueDate) {
-    this.dueDate = dueDate;
+    this.dueDate = new Date(dueDate.getTime());
   }
 
   public void setGrossTotal(Double grossTotal) {
